Apply consistent code style as used in other tests. Extract tests for App\Actions\SilentDevScript into a separate class.
